refactor(cli): extract the console registration suite into console-check.ts
check.ts had grown to 615 lines, 255 of them the console command
registration checks — a self-contained suite sitting inline while its
siblings (arch-check, docs-check, spec-check) each own a module that
check.ts merely invokes. Move it to console-check.ts to match that
shape: check.ts drops to ~350 lines of core checks plus orchestration,
and make:command now imports registersCommandsOf from the module that
owns it rather than from the aggregator.

moduleNameFor (absolute-path wrapper over moduleNameFromRelPath) moves
to discovery.ts, which already declares itself the single source of
truth for module-path-to-name, since both files need it. Tests move
with the code they exercise; no behavior change, all 3094 tests pass
unchanged.
